requirements.txt
Django==5.0.3
django-environ==0.9.0
google-cloud-storage==2.5.0
openai==0.22.0
